+++ web/src/views/Snooze.vue
@@ -6,15 +6,36 @@
       :form="form"
       :fields="fields"
       :tabs="tabs"
+      @row-selected="select"
       ref="table"
       edit_mode
       delete_mode
       add_mode
     >
-        <template v-slot:cell(hits)="row">
-          <router-link :to="get_link(dig(row.item, 'name'))">{{ dig(row.item, 'hits') }}</router-link>
-        </template>
+      <template v-slot:cell(hits)="row">
+        <router-link :to="get_link(dig(row.item, 'name'))">{{ dig(row.item, 'hits') }}</router-link>
+      </template>
+      <template #button="row">
+        <b-button variant="info" @click="modal_show([row.item], 'apply')" size="sm" v-b-tooltip.hover title="Retro apply"><i class="la la-redo la-lg"></i></b-button>
+      </template>
+      <template #selected_buttons>
+        <b-button variant="info" @click="modal_show(selected, 'apply')">Retro apply</b-button>
+      </template>
     </List>
+
+    <b-modal
+      id="modal"
+      ref="modal"
+      @ok="retro_apply(modal_data)"
+      @hidden="modal_clear()"
+      :header-bg-variant="modal_bg_variant"
+      :header-text-variant="modal_text_variant"
+      size="xl"
+      centered
+    >
+      <template #modal-title>{{ modal_title }}</template>
+      <p>{{ modal_message }}</p>
+    </b-modal>
   </div>
 </template>
 
@@ -24,7 +45,9 @@ import dig from 'object-dig'
 
 import List from '@/components/List.vue'
 
+import { text_alert } from '@/utils/query'
 import { form, fields, default_sorting } from '@/objects/Snooze.yaml'
+import { API } from '@/api'
 
 export default {
   components: {
@@ -35,6 +58,13 @@ export default {
   },
   data () {
     return {
+      modal_title: '',
+      modal_message: '',
+      modal_type: '',
+      modal_bg_variant: '',
+      modal_text_variant: '',
+      modal_data: [],
+      selected: [],
       dig: dig,
       form: form,
       fields: fields,
@@ -79,7 +109,47 @@ export default {
           s: encodeURIComponent(`snoozed=${escaped_name}`),
         },
       }
-    }, 
+    },
+    retro_apply(rows) {
+      var filter_names = []
+      rows.forEach(item => { filter_names.push(item.name) })
+      console.log(`Retro apply ${filter_names}`)
+      API
+        .put(`/snooze_apply`, filter_names)
+        .then(response => {
+          console.log(response)
+          if (response !== undefined && response.data) {
+            text_alert(`Updated ${response.data} alert(s)`, 'Success', 'success')
+          } else {
+            text_alert('An error occured', 'Failure', 'danger')
+          }
+        })
+        .catch(error => console.log(error))
+    },
+    select(items) {
+      this.selected = items
+    },
+    modal_clear() {
+      this.modal_data = []
+      this.modal_title = ''
+      this.modal_message = ''
+      this.modal_type = ''
+      this.modal_bg_variant = ''
+      this.modal_text_variant = ''
+    },
+    modal_show(items, type) {
+      this.modal_data = items
+      this.modal_type = type
+      this.modifications = []
+      switch (type) {
+        default:
+          this.modal_title = 'Retro apply to all alerts'
+          this.modal_message = 'This operation cannot be undone. Are you sure?'
+          this.modal_bg_variant = 'info'
+          this.modal_text_variant = 'white'
+      }
+      this.$bvModal.show('modal')
+    },
   },
 }
 </script>